ConfirmationMgr-ize `PtYesNoDialog()` python bindings.
This also includes the first stab at some general code that allows
firing an arbitrary callback from the engine into Python code without
leaking resources.

+++ Sources/Plasma/FeatureLib/pfPython/pyGlueHelpers.h
@@ -614,7 +614,7 @@ static PyObject *methodName(PyObject *self) /* and now for the actual function *
 #define PYTHON_ENUM_START(enumName) std::vector<std::tuple<ST::string, Py_ssize_t>> enumName##_enumValues{
 
 // for each element of the enum
-#define PYTHON_ENUM_ELEMENT(enumName, elementName, elementValue) std::make_tuple(ST_LITERAL(#elementName), elementValue),
+#define PYTHON_ENUM_ELEMENT(enumName, elementName, elementValue) std::make_tuple(ST_LITERAL(#elementName), (Py_ssize_t)elementValue),
 
 // to finish off and define the enum
 #define PYTHON_ENUM_END(m, enumName) }; pyEnum::MakeEnum(m, #enumName, enumName##_enumValues);
